+++ src/test/java/org/folio/mr/domain/mapper/MediatedBatchRequestMapperTest.java
@@ -28,7 +28,9 @@ class MediatedBatchRequestMapperTest {
   @Test
   void shouldMapPostDtoToEntity() {
     var batchId = UUID.randomUUID();
-    var postDto = new MediatedBatchRequestPostDto().batchId(batchId.toString());
+    var postDto = new MediatedBatchRequestPostDto()
+      .batchId(batchId.toString())
+      .mediatedWorkflow(MediatedBatchRequestPostDto.MediatedWorkflowEnum.MULTI_ITEM_REQUEST);
     var entity = mapper.mapPostDtoToEntity(postDto);
 
     assertEquals(batchId, entity.getId());
@@ -60,6 +62,7 @@ class MediatedBatchRequestMapperTest {
     entity.setUpdatedDate(Timestamp.from(Instant.now()));
     entity.setUpdatedByUserId(userId);
     entity.setUpdatedByUsername(username);
+    entity.setMediatedWorkflow("Multi-Item request");
 
     var dto = mapper.toDto(entity);